# lib/api-client-react/src/practice-history/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CCNA Learning SaaS API foundation
 * OpenAPI spec version: 0.2.0
 */
export interface PracticeHistoryItem {
  id: string;
  questionId: string;
  topicId: string;
  selectedOptionKey: string;
  isCorrect: boolean;
  submittedAt: string;
  /** Present only with includeDetails=true. Current question code; null if unavailable. */
  questionCode?: string | null;
  /** Present only with includeDetails=true. Current name of the topic recorded in the attempt, not the question's current topic; null if unavailable. */
  topicName?: string | null;
}

export interface PracticeHistoryResponse {
  items: PracticeHistoryItem[];
  nextCursor: string | null;
}

export type ErrorResponseError = {
  code: string;
  message: string;
};

export interface ErrorResponse {
  error: ErrorResponseError;
}

export type GetPracticeHistoryParams = {
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
/**
 * Opaque pagination cursor returned by a previous response.
 */
cursor?: string;
/**
 * Include nullable questionCode and topicName display fields. Omitted or false preserves the original response.
 */
includeDetails?: boolean;
};